+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/frame/DefaultFrameBuilder.java
@@ -20,8 +20,8 @@ package is.codion.swing.common.ui.frame;
 
 import is.codion.common.observable.Observable;
 import is.codion.common.value.Value;
-import is.codion.swing.common.ui.Windows;
 import is.codion.swing.common.ui.layout.Layouts;
+import is.codion.swing.common.ui.window.Windows;
 
 import javax.swing.ImageIcon;
 import javax.swing.JComponent;